+++ Src/DasherCore/TwoButtonDynamicFilter.cpp
@@ -121,7 +121,7 @@ void CTwoButtonDynamicFilter::KeyDown(int iTime, int iId, CDasherModel *pModel, 
     while((m_deQueueTimes.size() > 0) && (iTime - m_deQueueTimes.front()) > GetLongParameter(LP_HOLD_TIME))
       m_deQueueTimes.pop_front();
 
-    if(m_deQueueTimes.size() == GetLongParameter(LP_MULTIPRESS_COUNT) - 1) { 
+    if(m_deQueueTimes.size() == static_cast<unsigned int>(GetLongParameter(LP_MULTIPRESS_COUNT) - 1)) { 
       Event(iTime, iId, 2, pModel, pUserLog);
       m_deQueueTimes.clear();
     }
